requirements/base.txt
argon2-cffi==20.1.0  # https://github.com/hynek/argon2_cffi

# Django
# ------------------------------------------------------------------------------
django==3.0.10  # pyup: < 3.1  # https://www.djangoproject.com/
django-allauth==0.42.0  # https://github.com/pennersr/django-allauth
django-cors-headers==3.5.0 # https://github.com/adamchainz/django-cors-headers
django-crispy-forms==1.9.2  # https://github.com/django-crispy-forms/django-crispy-forms
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==4.0.0  # https://github.com/jazzband/django-model-utils
django-redis==4.12.1  # https://github.com/jazzband/django-redis
# Django REST Framework
djangorestframework==3.11.1  # https://github.com/encode/django-rest-framework
hiredis==1.1.0  # https://github.com/redis/hiredis-py
Pillow==7.2.0  # https://github.com/python-pillow/Pillow
python-slugify==4.0.1  # https://github.com/un33k/python-slugify
pytz==2020.1  # https://github.com/stub42/pytz
redis==3.5.3  # https://github.com/andymccurdy/redis-py
whitenoise==5.2.0  # https://github.com/evansd/whitenoise
django-rest-auth[with_social]==0.9.5 # https://github.com/Tivix/django-rest-auth
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.1 # https://github.com/axnsan12/drf-yasg#usage
